requirements.txt
alembic==1.14.0
annotated-types==0.7.0
anyio==4.7.0
asttokens==3.0.0
attrs==24.3.0
Automat==24.8.1
autopep8==2.3.1
bcrypt==4.2.1
cachetools==5.5.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.8
colorama==0.4.6
comm==0.2.2
constantly==23.10.4
cryptography==44.0.0
cssselect==1.2.0
debugpy==1.8.11
decorator==5.1.1
defusedxml==0.7.1
dnspython==2.7.0
ecdsa==0.19.0
email_validator==2.2.0
executing==2.1.0
fastapi==0.115.6
fastapi-cli==0.0.7
filelock==3.16.1
google-auth==2.37.0
google-auth-oauthlib==1.2.1
greenlet==3.1.1
gspread==6.1.4
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
ipykernel==6.29.5
ipython==8.31.0
itemadapter==0.10.0
itemloaders==1.3.2
jedi==0.19.2
Jinja2==3.1.5
jmespath==1.0.1
jupyter_client==8.6.3
jupyter_core==5.7.2
lxml==5.3.0
Mako==1.3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
mdurl==0.1.2
nest-asyncio==1.6.0
oauthlib==3.2.2
outcome==1.3.0.post0
packaging==24.2
parsel==1.9.1
parso==0.8.4
passlib==1.7.4
platformdirs==4.3.6
prompt_toolkit==3.0.48
Protego==0.3.1
psutil==6.1.1
psycopg2==2.9.10
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycodestyle==2.12.1
pycparser==2.22
pydantic==2.10.4
pydantic-settings==2.7.1
pydantic_core==2.27.2
PyDispatcher==2.0.7
Pygments==2.18.0
PyJWT==2.10.1
pyOpenSSL==24.3.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.20
pywin32==308
PyYAML==6.0.2
pyzmq==26.2.0
queuelib==1.7.0
requests==2.32.3
requests-file==2.1.0
requests-oauthlib==2.0.0
rich==13.9.4
rich-toolkit==0.12.0
rsa==4.9
Scrapy==2.12.0
scrapy-selenium==0.0.7
selenium==4.27.1
service-identity==24.2.0
setuptools==75.6.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
SQLAlchemy==2.0.36
sqlmodel==0.0.22
stack-data==0.6.3
starlette==0.41.3
tldextract==5.1.3
tornado==6.4.2
traitlets==5.14.3
trio==0.27.0
trio-websocket==0.11.1
Twisted==24.11.0
typer==0.15.1
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.34.0
w3lib==2.2.1
watchfiles==1.0.3
wcwidth==0.2.13
websocket-client==1.8.0
websockets==14.1
wsproto==1.2.0
XlsxWriter==3.2.0
zope.interface==7.2